# docs/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    RoleAssignmentOut:
      description: 'A grant of a role to a console account.


        `granted_at` is the assignment''s creation time; `updated_at` backs the

        `updated_since` reconciliation feed. Both the user and role are embedded so

        an IGA connector can reconcile without extra lookups.'
      properties:
        granted_at:
          description: When the role was granted (provisioned).
          format: date-time
          title: Granted At
          type: string
        id:
          title: Id
          type: integer
        role:
          $ref: '#/components/schemas/RoleAssignmentRoleRef'
        updated_at:
          format: date-time
          title: Updated At
          type: string
        user:
          $ref: '#/components/schemas/RoleAssignmentUserRef'
      required:
      - id
      - user
      - role
      - granted_at
      - updated_at
      title: RoleAssignmentOut
      type: object
    RoleAssignmentRoleRef:
      description: The role side of an assignment, embedded in assignment responses.
      properties:
        id:
          title: Id
          type: integer
        is_active:
          title: Is Active
          type: boolean
        name:
          title: Name
          type: string
      required:
      - id
      - name
      - is_active
      title: RoleAssignmentRoleRef
      type: object
    RoleAssignmentUserRef:
      description: The console-account side of an assignment, embedded in responses.
      properties:
        id:
          title: Id
          type: integer
        username:
          title: Username
          type: string
      required:
      - id
      - username
      title: RoleAssignmentUserRef
      type: object
    RoleCreate:
      description: Request body for creating an access role.
      properties:
        description:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Description
        is_active:
          default: true
          title: Is Active
          type: boolean
        name:
          maxLength: 100
          minLength: 1
          title: Name
          type: string
      required:
      - name
      title: RoleCreate
      type: object
    RoleGrant:
      description: Request body to grant (provision) a role to a console account.
      properties:
        role_id:
          description: ID of the role to grant.
          title: Role Id
          type: integer
      required:
      - role_id
      title: RoleGrant
      type: object
    RoleOut:
      description: An access role as returned by the roles API.
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          title: Id
          type: integer
        is_active:
          title: Is Active
          type: boolean
        name:
          title: Name
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - id
      - name
      - description
      - is_active
      - created_at
      - updated_at
      title: RoleOut
      type: object
    RoleUpdate:
      description: Request body for updating an access role. All fields optional.
      properties:
        description:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Description
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        name:
          anyOf:
          - maxLength: 100
            minLength: 1
            type: string
          - type: 'null'
          title: Name
      title: RoleUpdate
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Lightweight HR Source of Truth for Saviynt Identity Cloud POC and integration testing.
  title: Demo HR Source of Truth App
  version: 1.7.2
openapi: 3.1.0
paths:
...
  /api/v1/roles/:
    get:
      description: List the access-role catalog.
      operationId: list_roles_api_v1_roles__get
      parameters:
      - in: query
        name: is_active
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleOut'
                title: Response List Roles Api V1 Roles  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Roles
      tags:
      - roles
    post:
      description: Create an access role.
      operationId: create_role_api_v1_roles__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Role
      tags:
      - roles
  /api/v1/roles/assignments:
    get:
      description: 'Flat feed of every role grant, for pull-based reconciliation.


        Declared before ``/{role_id}`` so the literal path wins over the dynamic

        integer segment.'
      operationId: list_role_assignments_api_v1_roles_assignments_get
      parameters:
      - description: Only assignments for this console account.
        in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Only assignments for this console account.
          title: User Id
      - description: Only assignments of this role.
        in: query
        name: role_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Only assignments of this role.
          title: Role Id
      - description: ISO-8601 datetime. Returns only assignments granted or changed at or after this time
          — for incremental IGA reconciliation.
        in: query
        name: updated_since
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: ISO-8601 datetime. Returns only assignments granted or changed at or after this
            time — for incremental IGA reconciliation.
          title: Updated Since
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleAssignmentOut'
                title: Response List Role Assignments Api V1 Roles Assignments Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Role Assignments
      tags:
      - roles
  /api/v1/roles/{role_id}:
    delete:
      description: 'Delete an access role.


        Blocked while the role is still granted to any account — deactivate it

        (``is_active=false``) or revoke the assignments first.'
      operationId: delete_role_api_v1_roles__role_id__delete
      parameters:
      - in: path
        name: role_id
        required: true
        schema:
          title: Role Id
          type: integer
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Role
      tags:
      - roles
    get:
      description: Get a single access role by ID.
      operationId: get_role_api_v1_roles__role_id__get
      parameters:
      - in: path
        name: role_id
        required: true
        schema:
          title: Role Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Role
      tags:
      - roles
    patch:
      description: Update an access role.
      operationId: update_role_api_v1_roles__role_id__patch
      parameters:
      - in: path
        name: role_id
        required: true
        schema:
          title: Role Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Role
      tags:
      - roles
...
  /api/v1/users/{user_id}/roles:
    get:
      description: List the access roles currently granted to a console account.
      operationId: list_user_roles_api_v1_users__user_id__roles_get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleAssignmentOut'
                title: Response List User Roles Api V1 Users  User Id  Roles Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List User Roles
      tags:
      - users
    post:
      description: Grant (provision) an access role to a console account.
      operationId: grant_user_role_api_v1_users__user_id__roles_post
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleGrant'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignmentOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Grant User Role
      tags:
      - users
  /api/v1/users/{user_id}/roles/{role_id}:
    delete:
      description: Revoke (deprovision) an access role from a console account.
      operationId: revoke_user_role_api_v1_users__user_id__roles__role_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      - in: path
        name: role_id
        required: true
        schema:
          title: Role Id
          type: integer
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke User Role
      tags:
      - users
...